requirements.txt
streamlit>=1.32.0
pandas>=2.0.0,<3.0
supabase>=2.10.0,<3.0
python-dotenv>=1.0.0
pytest>=8.0.0
requests>=2.31.0
